Add map and popup panel that we have been using for OpenLayers.
This adds in the old way we were doing things with OpenLayers.
A slight variation of this has been used in production applications but
we are going to be replacing it with the new google maps method.
I am committing it here so we can keep it for reference and so
it can be of use to other people that may want something similar.