Refactor the navigation control overrides to use selectors.
The old code made use of helper functions to create items
with specific itemIds so we could get automatically created
navigation UI widgets.

This change replaces this method with a more general use of
selectors.  This means users have to specify the UI in their
own code, but can use selectors to get the stack controller
to help with naming, etc.